requirements.txt
django==3.2.16
django-simpleui==2022.11.30
user_agents==2.2.0
requests==2.28.1
beautifulsoup4==4.11.1
faker==15.3.4
lxml==4.9.1
django-ckeditor==6.5.1
django_notifications_hq==1.7.0
nanoid==2.0.0